# doc/part-3/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Volunteer Information System API
  description: API для системы управления волонтерскими мероприятиями
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          scopes:
            read: Чтение данных
            write: Запись данных
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор пользователя
        organisationId:
          type: integer
          nullable: true
          description: ID организации (для представителей организации)
        role:
          type: string
          enum: [USER, ORG_REPRESENTATIVE, ADMIN]
          description: Роль пользователя
        name:
          type: string
          description: Полное имя пользователя
        email:
          type: string
          format: email
          description: Email пользователя
        description:
          type: string
          nullable: true
          description: Описание "о себе"
        birthday:
          type: string
          format: date
          nullable: true
          description: День рождения
        location:
          type: string
          nullable: true
          description: Место проживания
        volunteerHours:
          type: number
          format: float
          default: 0
          description: Количество отработанных волонтерских часов
        rating:
          type: number
          format: float
          nullable: true
          description: Средний рейтинг пользователя по отзывам
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - fullName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT токен для авторизации
        user:
          $ref: '#/components/schemas/User'
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      description: Регистрация нового пользователя через email и пароль
      requestBody: 
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [auth]
      description: Вход пользователя через email и пароль
      requestBody: 
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Пользователь успешно вошел
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/oauth2/{provider}/authorize:
    get:
      tags: [auth]
      summary: OAuth2 авторизация
      description: Перенаправляет пользователя на страницу провайдера (VK, Yandex)
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [vk, yandex]
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
          description: URL для возврата после авторизации
      responses:
        '302':
          description: Перенаправление на страницу провайдера
        '400':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/oauth2/{provider}/callback:
    get:
      tags: [auth]
      summary: OAuth2 callback
      description: Обрабатывает ответ от провайдера и возвращает токен аутентификации
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [vk, yandex]
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Код авторизации, полученный от провайдера
        - name: state
          in: query
          required: false
          schema:
            type: string
          description: Состояние для предотвращения CSRF атак
      responses:
        '200':
          description: Пользователь успешно аутентифицирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags: [auth]
      summary: Выход пользователя
      description: Завершает сеанс пользователя, инвалидируя токен
      responses:
        '200':
          description: Пользователь успешно вышел
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # User Profile
...